ollama/ollama-audio/requirements.txt
--extra-index-url='https://download.pytorch.org/whl/cu128' # channge if necessary

pyfiglet
ollama
pillow
tts
huggingface